Change extensions and Prettier formatting
