--- src/Playlist/Playlist.js
@@ -1,9 +1,0 @@
-import styles from "./Playlist.module.css";
-
-export default function Playlist(){
-    return (
-        <>
-        
-        </>
-    )
-}
