+++ src/Playlist/Playlist.jsx
@@ -1,0 +1,5 @@
+import styles from "./Playlist.module.css";
+
+export default function Playlist() {
+  return <></>;
+}
