--- src/SearchResults/SearchResults.js
@@ -1,9 +1,0 @@
-import styles from "./SearchResults.module.css";
-
-export default function SearchResults(){
-    
-    return(
-        <>
-        </>
-    )
-}
