+++ src/SearchResults/SearchResults.jsx
@@ -1,0 +1,5 @@
+import styles from "./SearchResults.module.css";
+
+export default function SearchResults() {
+  return <></>;
+}
