--- src/Track/Track.js
@@ -1,12 +1,0 @@
-import styles from "./Track.module.css";
-
-export default function Track(){
-
-    return(
-        <>
-            <form>
-                <button type="submit">Save To Spotify</button>
-            </form>
-        </>
-    )
-}
