+++ src/Track/Track.jsx
@@ -1,0 +1,11 @@
+import styles from "./Track.module.css";
+
+export default function Track() {
+  return (
+    <>
+      <form>
+        <button type="submit">Save To Spotify</button>
+      </form>
+    </>
+  );
+}
